+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './operations';
 export * from './types';
 export { Store } from './store';
-export { cacheExchange } from './exchange';
-export { populateExchange } from './populate';
+export { cacheExchange } from './cacheExchange';
+export { populateExchange } from './populateExchange';
